requirements.txt
beautifulsoup4==4.14.2
GitPython==3.1.45
Markdown==3.5.1
Office365-REST-Python-Client==2.5.12
pandas==2.1.4
PGPy==0.6.0
pycryptodome==3.23.0
pycryptodomex==3.23.0
pyotp==2.9.0
requests==2.32.4
requests-cache==1.2.1
scikit-learn==1.7.2
tabulate==0.9.0
Whoosh==2.7.4
